requirements.txt
fastapi
uvicorn
openai
pydantic
whisper
speechrecognition
numpy
spacy
python-dotenv
soundfile